refactor: rename client package to sse
